+++ modules/@apostrophecms/polymorphic-type/index.js
@@ -1,10 +1,19 @@
 const _ = require('lodash');
+const migrations = require('./lib/migrations.js');
 
 module.exports = {
   extend: '@apostrophecms/doc-type',
   options: {
-    name: '@apostrophecms/polymorphic',
+    name: '@apostrophecms/polymorphic-type',
     showPermissions: false
+  },
+  init(self) {
+    self.addMigrations();
+  },
+  methods(self) {
+    return {
+      ...migrations(self)
+    };
   },
   routes(self) {
     return {